mac80211: handle drv_add_interface failures properly during reconfig
If any interface fails to be added to the driver in during reconfig,
we should remove all the successfully added interfaces and report
reconfig failure, so things can be cleaned up properly.  Failing to do
so can lead to subsequent failures and leave the drivers in a messed
up state.

Change-Id: If2baff65bcd50f73c7f30fbd3c502326c63223c2
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/51816
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/52075
Tested-by: Johannes Berg <johannes.berg@intel.com>
x-iwlwifi-stack-dev: 5f3719247b188f7632486935f56ce2defaa96d0c