iwlwifi: mvm: don't try to stop scans that are not running anymore
In certain conditions, mac80211 may ask us to stop a scan (scheduled
or normal) that is not running anymore.  This can also happen when we
are doing a different type of scan, for instance, mac80211 can ask us
to stop a scheduled scan when we are running a normal scan, due to
some race conditions.  In this case, we would stop the wrong type of
scan and leave everything everything in a wrong state.

To fix this, simply ignore scan stop requests for scans types that are
not running.

Change-Id: I3f765a316ae947af448606b0b37903e52f429449
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
---
MWG100228484
Reviewed-on: https://gerrit.rds.intel.com/r/52108
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins
Reviewed-on: https://gerrit.rds.intel.com/r/52110
Tested-by: Johannes Berg <johannes.berg@intel.com>
x-iwlwifi-stack-dev: 8782b76a03b7e28da79c7da6d818a2b9fc6a85af