trans_sdio: Remove the trigger flags when requesting the OOB interrupt.
From the hw spec of iwl devices, the OOB line is pulled up all the time,
and is driven by the chipset when an interrupt is raised to the host.
Usually a Falling edge trigger for the OOB interrupt covers the needs.
But a better implementation is to rely on the platform's firmware
(BIOS on x86, Open Firmware, ...) for the trigger and the polarity.

Change-Id: I74492eae17c8b45578c08d13fc9c5d1fc0d524ef
Signed-off-by: Aymen Zayet <aymen.zayet@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/52115
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-by: Mariane Ravet <mariane.ravet@intel.com>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/52179
Tested-by: IWL Jenkins
x-iwlwifi-stack-dev: eaff5483f4f7c2e99b5310bb1697621b552e995a