mac80211: fix another suspend vs. association race
Since cfg80211 disconnects, but has no insight into the association
process, it can happen that it disconnects while association is in
progress. We then try to abort association in mac80211, but this is
only later so the association can complete between the two.

This results in removing an interface from the driver while bound
to the channel context, obviously causing confusion and issues.

Solve this by also checking if we're associated during quiesce and
if so deauthenticating. The frame will no longer go out to the AP
which is a bit unfortunate, but it'll resolve the crash (and before
we would have suspended without telling the AP as well.)

I'm working on a better, but more complex solution as well, which
should avoid that problem.

Change-Id: I5dc760afb102bdb61eb1147273e8e0eb08cb6ccf
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
---
This fixes another issue seen in chrome-os-partner:34821
but is not the root cause for these issues.
Reviewed-on: https://gerrit.rds.intel.com/r/52131
Reviewed-on: https://gerrit.rds.intel.com/r/52133
x-iwlwifi-stack-dev: 7c06cbe3b5a2348c1b8c689c620ebf18470f7910