iwlwifi: mvm: Fix ROC removal [BUGFIX]
iwl_mvm_stop_roc removes TE only if running flag is set. This is not correct
since this flag is only set when the TE is started.
This resulted in a TE not being removed, when mac80211 believes that there are
no active ROCs.
---
MWG100227155
MWG100227540

This was introduced in
0ed058a85 "iwlwifi: mvm: add remove flow for AUX ROC time events"

Change-Id: I550a1b028d928d1e664a33bb33d54671b9852d7c
Signed-off-by: Andrei Otcheretianski <andrei.otcheretianski@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/52514
Reviewed-by: David Spinadel <david.spinadel@intel.com>
Tested-by: IWL Jenkins
Reviewed-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-by: Matti Gottlieb <matti.gottlieb@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/52624
x-iwlwifi-stack-dev: 83a87182086b9600c0707b38fee5623b731908af