iwlwifi: mvm: fix logic errors in U-APSD non-agg detection
There are two logic errors - one concerns the use of high bandwidths,
where the expected throughput has to be multiplied by 1, 2, 4 or 8
instead of 1, 2, 3 or 4 (for 20, 40, 80 and 160 MHz respectively).

The second concerns the throughput calculation - to compare against
the expected throughput we need the same unit (100 Kbps), which was
erroneous as 10*8*bytes/msec is 10,000 bps = 10 Kbps, not 100 Kbps
as the expected (average) rate.

While at it also use the correct unit prefix (Kbps.)

Change-Id: I0b705b73d2896e0c3dfc65cb37d252f99fbcaa82
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/52996
Tested-by: IWL Jenkins
Reviewed-on: https://gerrit.rds.intel.com/r/53111
x-iwlwifi-stack-dev: 6fbc0816c45fb8b6bef9e7b516585c2a20ccd797