[NOUPSTREAM] iwlwifi: mvm: implement device TX power using vendor command
As the device TX power limit doesn't really reflect well in the
upstream API, and the GO SRD limit conflicts with it, requires
hacks to implement properly and will be hard to maintain, use a
vendor command for now to set the device TX power. This will be
used for SAR limits and the supplicant will provide the API for
real usage, which allows it to be adjusted again later.

This addresses the conflict between SAR and SRD limits.

type=bugfix
bug=cq:MWG100229536

Change-Id: I95bfea3bd558f40855fba8f50d97466606b7397d
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/53419
Reviewed-by: Ilan Peer <ilan.peer@intel.com>
Tested-by: IWL Jenkins
Reviewed-on: https://gerrit.rds.intel.com/r/53775
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 883ed8ddd5124c06a6af536b6bfb2f671831ae38