[BUGFIX] mac80211: ask for ECSA IE to be considered for beacon parse CRC
When a beacon from the AP contains only the ECSA IE, and not a CSA IE
as well, this ECSA IE is not considered for calculating the CRC and
the beacon might be dropped as not being interesting. This is clearly
wrong, it should be handled and the channel switch should be executed.

Fix this by including the ECSA IE ID in the bitmap of interesting IEs.

type=bugfix
bug=MWG100230965

Change-Id: Ia6a61574351dbcb7bc7d659d474ffeadb01a055f
Reported-by: Gil Tribush <gil.tribush@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/53666
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-by: Luciano Coelho <luciano.coelho@intel.com>
Tested-by: IWL Jenkins
Reviewed-on: https://gerrit.rds.intel.com/r/53933
x-iwlwifi-stack-dev: 9a3e63aef4fd5c3be448982c6b5fc20747123859