mac80211: ignore CSA to same channel
If the AP is confused and starts doing a CSA to the same channel,
just ignore that request instead of trying to act it out since it
was likely sent in error anyway.

In the case of the bug I was investigating the GO was misbehaving
and sending out a beacon with CSA IEs still included after having
actually done the channel switch.

type=bugfix
bug=cq:MWG100230962

Change-Id: Ic6c388469440143c0e7b827645e3f2dba3779d25
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/53736
Reviewed-by: Luciano Coelho <luciano.coelho@intel.com>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Tested-by: IWL Jenkins
Reviewed-on: https://gerrit.rds.intel.com/r/53779
x-iwlwifi-stack-dev: 4f2901b823af858345e8011dc6fc3497fa9b35ea