iwlwifi: trans_slave: implement wait_for_empty
This allows the op_mode to make sure that the transport
queues are empty. This will be needed to flush the queues
properly before removing them.
Note that the op_mode will wait for the queues to be empty
also in disconnection / roaming flows.

Change-Id: I8e5065dbfd91c4fd3b65ee759247d16582136c5b
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/53803
Tested-by: IWL Jenkins
Reviewed-on: https://gerrit.rds.intel.com/r/54053
x-iwlwifi-stack-dev: 0da5f0aad14ebcff89effc780b92e04b59560b96