iwlwifi: mvm: properly flush the queues for buffering transport [BUGFIX]
There are transport that must buffer frames in the driver.
This means that we have frames that are not in the op_mode
and not visible to the firwmare. This causes issues when we
flush the queues: the op_mode flushes a queue, and the
firmware flushes all the frames that are *currently* on the
rings, but if the transport buffers frames, it can submit
these while we are flushing. This leads to a situation
where we still have frames on the queues after we flushed
them.
Preventing those buffered frame from getting into the
firmware is possible, but then, we have to run the Tx
response path on frames that didn't reach the firmware
which is not desirable.
The way I solve this here is to let these frames go to the
firmware, but make sure the firmware will not transmit them
(by setting the station as draining). The op_mode then needs
to wait until the transport itself is empty to be sure that
the queue is really empty.

Change-Id: I4475db262d82283555d2d99ec5fe93bf60b3b8fd
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
---
MWG100231132
Reviewed-on: https://gerrit.rds.intel.com/r/53804
Tested-by: IWL Jenkins
Reviewed-on: https://gerrit.rds.intel.com/r/54054
x-iwlwifi-stack-dev: d51326a90122533a078d69813137701929a46695