mac80211: Disable U-APSD if connected to a SISO AP
This concerns a bugy behavior of some APs, that advertise themselves
falsely as supporting U-APSD, but they don't which affect throughput.
It was detected in iPhones, but recently also with some Netgear models.
Those devices also advertise their capabilities as SISO, so use that
and disable U-APSD if connected to a SISO AP.

Change-Id: I1271e3b86e29e8f3efce1ebf725281790388d7bd
Signed-off-by: Avri Altman <avri.altman@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/53464
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins
Reviewed-on: https://gerrit.rds.intel.com/r/57862
Reviewed-by: Guy Cohen <guy.cohen@intel.com>
x-iwlwifi-stack-dev: ed6ec9bd46ef0f5572079430044d6c3200026a85