iwlwifi: mvm: always reconfigure last MCC on init [bugfix]
Currently the last found MCC is reconfigured only in the recovery flow.
But it should always be used when available, for the ifdown/up or
RF-Kill/CT-Kill scenarios.
While at it, fix a couple of bugs in the init-from-last-MCC flow. Return
an error value when a current MCC is not found. Pass on the regdomain to
cfg80211 only if it was changed and don't ignore the return value from
the cfg80211-setter function.

Change-Id: Ic184e5d749e5648e15424f5cdcd519b2930f7de5
Signed-off-by: Arik Nemtsov <arikx.nemtsov@intel.com>
--
MWG100232023
Reviewed-on: https://gerrit.rds.intel.com/r/54419
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/54797
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 6a2623f05e45a7806a4a166aa0128ec445867138