[BUGFIX] iwlwifi: mvm: rs: report last tx rate based on RSSI and caps
In scenarios where we haven't converged yet to a specific modulation
and rate it could be better to report to userspace the last tx rate
based on the STA capabilities and RSSI. This is important as sometimes
userspace displays the last tx rate as the link speed.
This avoids being presented with low legacy rates when rs just begins
its search or after an idle period in which it resets itself.

type=bugfix
bug=cq:MWG100234428
fixes=unknown

Change-Id: I78c428b2b40c852188cd6efea68d07b064b811bb
Signed-off-by: Eyal Shapira <eyalx.shapira@intel.com>
---
This better matches Ayelet's Link Indications SAS
Marked as fixed=unknown as this wasn't something that was broken
by a specific commit but was there all along.
Reviewed-on: https://gerrit.rds.intel.com/r/57733
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/58012
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 974b4d7351f190f7cdc3b52d8ae7f317d7b8e640