[BUGFIX] mac80211: ignore invalid scan RSSI values
Channels in 2.4GHz band overlap, this means that if we
send a probe request on channel 1 and then move to channel
2, we will hear the probe response on channel 2. In this
case, the RSSI will be lower than if we had heard it on
the channel on which it was sent (1 in this case).

The scan result ignores those invalid values and the
station last signal should not be updated as well.

In case the scan determines the signal to be invalid turn on
the flag so the station last signal will not be updated with
the value and thus user space probing for NL80211_STA_INFO_SIGNAL
and NL80211_STA_INFO_SIGNAL_AVG will not get this invalid RSSI
value.

type=bugfix
bug=chrome-os-partner:38941
fixes=unknown

Change-Id: I2775e8ffef46c37a2e0bccd509c02d4464790f90
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/57812
Tested-by: IWL Jenkins
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/58019
x-iwlwifi-stack-dev: fada403d59a1221ce55e7f71689910b5eda8b81b