[BUGFIX] iwlwifi: mvm: Avoid accessing Null pointer when setting igtk
Sometimes when setting an igtk key the station maybe NULL.
In the of case igtk the function will skip to the end, and
try to print sta->addr, if sta is Null - we will access a
Null pointer.

Avoid accessing a Null pointer when setting a igtk key &
the sta == NULL, and print a default MAC address instead.

type=bugfix
bug=not-tracked
fixes=unknown

Change-Id: I17b71856adfde53cb864d12d665e196e0cb24d23
Signed-off-by: Matti Gottlieb <matti.gottlieb@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/57945
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Tested-by: IWL Jenkins
Reviewed-on: https://gerrit.rds.intel.com/r/57986
x-iwlwifi-stack-dev: e38090134ae0105903ca1aae69edf8ceda8ccf50