mac80211: stop using pointers as userspace cookies
Even if the pointers are really only accessible to root and used
pretty much only by wpa_supplicant, this is still not great; even
for debugging it'd be easier to have something that's easier to
read and guaranteed to never get reused.

With the recent change to make mac80211 create an ack_skb for the
mgmt-tx path this becomes possible, only the client probe method
needs to also allocate an ack_skb, and we can store the cookie in
that skb.

type=feature

Change-Id: I341d735dcfa09c1ad4dd84d638639c4321986f49
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/57949
Tested-by: IWL Jenkins
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
---
As it turns out, this is actually a bugfix:
My commit "mac80211: copy nl80211 mgmt TX SKB for status" broke
status reporting because it used the copy of the SKB, but I only
adjusted the cookie sent to userspace in case of an on-channel
mgmt-tx, not in the case of an off-channel (with an implicit
remain-on-channel) transmission.
Reviewed-on: https://gerrit.rds.intel.com/r/58602
x-iwlwifi-stack-dev: 73bde00f21a4faeb3ee30b8c0c31ca9e4dd9cf11