[BUGFIX] mac80211: Disable U-APSD if connected to a non-HT AP
This concerns a buggy behavior of iPhone 4/4s when SoftAP, that advertise
themselves falsely as supporting U-APSD.  However, they don't raise the
TIM in its Beacon even if it has data for us, therefore traffic is stuck.
As a workaround, use the fact that those devices does not advertise HT
capabilities, and disable U-APSD if connected to a non-HT AP.
Our previous implementation assumed that if it doesn't have ht_capa
it can't be a SISO iphone.  Evidently this is not true, as non-HT implies
that the AP does not support 11n, hence does not support MIMO.

type=bugfix
bug=CQ:MWG100237205
fixes:I2ea5d8109e0e5ae8b0a24e3f78ff5d0b500022ea

Change-Id: Ica7c68d81849b5800887a6f0d7a8cc7d295fe77e
Signed-off-by: Avri Altman <avri.altman@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/58100
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Tested-by: IWL Jenkins
Reviewed-on: https://gerrit.rds.intel.com/r/58405
x-iwlwifi-stack-dev: c8651e041654151d4a4a0e50799ae9aa2845e2ba