[BUGFIX] iwlwifi: Android make: Solve multiple iwlwifi issue
In the CHT51 tree we may have multiple iwlwifi
in different folders, this will cause that the
only one that will be evaluated is the last one
included. the preferred way to handle this is to
include the right one from the BoardConfig.mk file,
but looks like some of the needed variables are
not defined when this file is included, what will
cause a failure in the build. so we need to include
the file later on, when all Android mk are included
and all variables are defined. so we do this from
our Android.mk

fixes=unknown
type=bugfix
bug=not-tracked

Signed-off-by: Benzi Zbit <benzi.zbit@intel.com>
Change-Id: If7bb9684c4bd1df5d1d8672933bab9a1dcb6a803
Reviewed-on: https://gerrit.rds.intel.com/r/58569
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Tested-by: IWL Jenkins
Reviewed-on: https://gerrit.rds.intel.com/r/58720
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: f139cb09534405e9eb7b215f3f382397c0302340