mac80211: fixup AIFSN instead of disabling WMM [BUGFIX]
Disabling WMM has a huge impact these days. It implies that
HT and VHT will be disabled which means that the throughput
will be drammatically reduced.
Since the AIFSN is a transmission parameter, we can play a
bit and fix it up to make it compliant with the 802.11
specification which requires it to be at least 2.
Increasing it from 1 to 2 will slightly reduce the
likelyhood to get a transmission opportunity compared to
other clients that would accept to set AIFSN=1, but at
least it will allow HT and VHT which is a huge gain.

type=bugfix
bug=cq:MWG100237827
fixes=unknown

Change-Id: I233a5e34ac5d6c80a6398618ddd0f2fbe9abb80b
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/58605
Reviewed-on: https://gerrit.rds.intel.com/r/58616
Tested-by: IWL Jenkins
x-iwlwifi-stack-dev: 18908cb336691d68bace7a6b4c242c14bc09ec99