[BUGFIX] iwlwifi: mvm: convert world-reg requests to get-def MCC
If a device has a default alpha2 configured in the NVM, the correct
response for a world-regdom request is the default alpha2 with its regdom.
Change the "00" request to "ZZ" for the FW to make this happen.

type=bugfix
bug=not-tracked
fixes=unknown

Change-Id: I1d9d2b0ff3a2665b432147e5d83aa50d17967a59
Signed-off-by: Arik Nemtsov <arikx.nemtsov@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/59607
Reviewed-by: Oren Haggai <oren.haggai@intel.com>
Tested-by: Oren Haggai <oren.haggai@intel.com>
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
x-iwlwifi-stack-dev: 498c4cbc256b6dcbfafebb3bf5a0bae629a343b8