iwlwifi: mvm: Fix UMAC target access read
Currently when reading data from UMAC memory sometimes the driver
get garbage. This happens since the driver is reading from an
invalid address.

The top 2 bits in the UMAC addresses are used for:
1. Telling the ARC MMU not to do address translation
2. Not to do cache access
When the driver trys to read using these bits it can lead to corrupt
access.

In order for the driver to access these address in needs to remove the 2
most significant bits.

Alter the UMACS base address of the error table so that the error dump
will be valid.

type=bugfix
bug=not-tracked
fixes=unknown

Change-Id: Ib796ba7a0db1c787a22a8f7a571dcb3688137ad1
Signed-off-by: Matti Gottlieb <matti.gottlieb@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/59578
Reviewed-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Tested-by: IWL Jenkins
Reviewed-by: Dor Shaish <dor.shaish@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/59911
x-iwlwifi-stack-dev: 0a484f752e50db7fffe2952baed6830052926404