iwlwifi: pcie: fix prepare card flow
When the card is not owned by the PCIe bus, we need to
acquire ownership first. This flow is implemented in
iwl_pcie_prepare_card_hw. Because of a hardware bug, we
need to disable link power management before we can
request ownership otherwise the other user of the device
won't get notified that we are requesting the device which
will prevent us from acquire ownership.

Same holds for the down flow where we need to make sure
that any other potential user is notified that the driver
is going down.

type=bugfix
bug=not-tracked
fixes=unknown

CC: <stable@vger.kernel.org> [4.1]
Change-Id: I6d7c1e719f7b1d142949140fe5770c52bd648f50
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/60312
Tested-by: IWL Jenkins
Reviewed-on: https://gerrit.rds.intel.com/r/60362
x-iwlwifi-stack-dev: 5520c92341e0d8262897e39acd984fcd315229b2