[BUGFIX] iwlwifi: pcie: don't synchronize_irq from the ISR
It is usually a bad idea to call synchronize_irq() from the
ISR. Even more so with a from the ISR and with mutex held.
Since rfkill is an interrupt and it can call
iwl_trans_stop_device, we can get to a situation where we
call synchronize_irq from the ISR with trans_pcie->mutex
held which prevents other threads from running.

For the moment, just remove the synchronize_irq in
iwl_trans_stop_device. We will need to find a solution for
the interrupts that might be waiting later.

type=bugfix
bug=cq:MWG100241792
fixes=I52c58956854c7c5a71345024091054e97545953a

Change-Id: I6bd5707310c28cfdaee4c26f210d0d37d878827e
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/60526
Tested-by: IWL Jenkins
Reviewed-on: https://gerrit.rds.intel.com/r/60554
x-iwlwifi-stack-dev: 6deb4657d269196131fd69d8d436e41de74d3178