iwlwifi: mvm: remove netdetect debugfs entry
This debugfs entry was implemented just as a proof-of-concept before
the full support for netdetect via cfg80211/mac80211 was implemented.
Now that we have a proper way to enable netdetect from userspace, this
entry is redundant and should be removed to avoid unnecessary
maintenance work.

type=cleanup

Change-Id: I4b163be2ff62f828ec11b2a0c822e8ddbf92cae8
Signed-off-by: Luciano Coelho <luciano.coelho@intel.com>
Reviewed-on: https://gerrit.rds.intel.com/r/61481
Reviewed-by: Johannes Berg <johannes.berg@intel.com>
Reviewed-by: Avraham Stern <avraham.stern@intel.com>
Tested-by: IWL Jenkins
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/3651
Tested-by: ec ger unix iil jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
x-iwlwifi-stack-dev: bee03ae7fcac7dcd0358bcbaf0254f15686d3524