[BUGFIX] iwlwifi: mvm: fix statistics variables type
When receiving statistics notification there is a field of average energy.
This is defines as signed 8 bit, while FW refers to it as unsigned.
when the energy is higher than 127 this causes in iwl_mvm_stat_iterator
a wrong computation of the signal
	int sig = -stats->general.beacon_filter_average_energy;
resulting in incorrect CQM event (change from low to high).

While at it - update the rest of the fields to the correct type.

type=bugfix
fixes=unknown
bug=cq:MWG100243046

Change-Id: I7d31dc901cd3254869b0a5b9d69f90fd24967ef6
Signed-off-by: Sara Sharon <sara.sharon@intel.com>
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/1954
Reviewed-by: Berg, Johannes <johannes.berg@intel.com>
Tested-by: Berg, Johannes <johannes.berg@intel.com>
x-iwlwifi-stack-dev: cf4d3c493fcad75f2cfcfd61b4a0a9cdb66e733c